chocolate-curry-11 Implement Item Sets - Refactor common types and interfaces into a Models folder
