+++ app/page.tsx
@@ -2,17 +2,10 @@ import {ThemeToggle} from "@/components/ThemeToggle";
 import TierListManager from "@/components/TierListManager";
 import {ZenToggle} from "@/components/ZenToggle";
 import imagesetConfig from '../imageset.config.json';
-import {ItemProps} from "@/components/Item";
+import Item from "@/models/Item";
+import Tier, {DEFAULT_TIER_TEMPLATE} from "@/models/Tier";
 
-export interface Tier {
-  id: string;
-  name: string;
-  items: ItemProps[];
-  labelPosition?: 'top' | 'left' | 'right';
-  placeholder?: string;
-}
-
-const getInitialItems = (): ItemProps[] => {
+const getInitialItems = (): Item[] => {
   return imagesetConfig.packageImages.flatMap(({packageName, images}) =>
     images.map((image: string, index: number) => ({
       id: `${packageName}-item-${index}`,
@@ -22,18 +15,12 @@ const getInitialItems = (): ItemProps[] => {
   );
 };
 
-const getInitialTiers = (items: ItemProps[]): Tier[] => [
-  {id: 'tier-s', name: 'S', items: [], labelPosition: 'left', placeholder: 'S'},
-  {id: 'tier-a', name: 'A', items: [], labelPosition: 'left', placeholder: 'A'},
-  {id: 'tier-b', name: 'B', items: [], labelPosition: 'left', placeholder: 'B'},
-  {id: 'tier-c', name: 'C', items: [], labelPosition: 'left', placeholder: 'C'},
-  {id: 'tier-f', name: 'F', items: [], labelPosition: 'left', placeholder: 'F'},
-  {id: 'uncategorized', name: '', items: items, labelPosition: 'left'},
-];
-
 const Home = () => {
   const initialItems = getInitialItems();
-  const initialTiers = getInitialTiers(initialItems);
+  const initialTiers: Tier[] = [
+    ...DEFAULT_TIER_TEMPLATE,
+    {id: 'uncategorized', name: '', items: initialItems, labelPosition: 'left'},
+  ];
 
   return (
     <main className="flex flex-col items-center justify-between">
