+++ components/TierTemplateSelector.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import {Tier} from "@/app/page";
 import {
   DropdownMenu,
   DropdownMenuContent,
@@ -12,33 +11,7 @@ import {
 import {Button} from "@/components/ui/button";
 import {LayersIcon} from "@radix-ui/react-icons";
 import {useTierContext} from '@/contexts/TierContext';
-
-type TierTemplate = Record<string, Tier[]>;
-export type LabelPosition = 'top' | 'left' | 'right';
-
-const tierTemplates: TierTemplate = {
-  '3rows': [
-    {id: 'tier-s', name: 'S', items: [], labelPosition: 'left', placeholder: 'S'},
-    {id: 'tier-a', name: 'A', items: [], labelPosition: 'left', placeholder: 'A'},
-    {id: 'tier-b', name: 'B', items: [], labelPosition: 'left', placeholder: 'B'},
-  ],
-  '5rows': [
-    {id: 'tier-s', name: 'S', items: [], labelPosition: 'left', placeholder: 'S'},
-    {id: 'tier-a', name: 'A', items: [], labelPosition: 'left', placeholder: 'A'},
-    {id: 'tier-b', name: 'B', items: [], labelPosition: 'left', placeholder: 'B'},
-    {id: 'tier-c', name: 'C', items: [], labelPosition: 'left', placeholder: 'C'},
-    {id: 'tier-f', name: 'F', items: [], labelPosition: 'left', placeholder: 'F'},
-  ],
-  '7rows': [
-    {id: 'tier-ss', name: 'SS', items: [], labelPosition: 'left', placeholder: 'SS'},
-    {id: 'tier-s', name: 'S', items: [], labelPosition: 'left', placeholder: 'S'},
-    {id: 'tier-a', name: 'A', items: [], labelPosition: 'left', placeholder: 'A'},
-    {id: 'tier-b', name: 'B', items: [], labelPosition: 'left', placeholder: 'B'},
-    {id: 'tier-c', name: 'C', items: [], labelPosition: 'left', placeholder: 'C'},
-    {id: 'tier-d', name: 'D', items: [], labelPosition: 'left', placeholder: 'D'},
-    {id: 'tier-f', name: 'F', items: [], labelPosition: 'left', placeholder: 'F'},
-  ],
-};
+import {LabelPosition, TierTemplate, tierTemplates} from "@/models/Tier";
 
 const TierTemplateSelector: React.FC = () => {
   const {
