+++ models/Item.ts
@@ -1,0 +1,7 @@
+export default interface Item {
+  id: string;
+  content: string;
+  imageUrl?: string;
+  onDelete?: (id: string) => void;
+  showLabel?: boolean;
+}
